Move cron code into a library
Amend camera scheduling to use the library
